+++ e2e/tests/media.e2e.js
@@ -140,6 +140,12 @@ describe('plugins', () => {
           }
         ).should('be.visible');
         cy.eyesCheckWindow({ tag: this.test.title, target: 'window', fully: false });
+        cy.get(
+          '#pgi65a6266ba23a8a55da3f469157f15237_0 > :nth-child(1) > .gallery-item-wrapper > :nth-child(1) > a > .gallery-item-content > .gallery-item-visible',
+          {
+            timeout: 10000,
+          }
+        ).should('be.visible');
         cy.get(`[data-hook=${'nav-arrow-next'}]`).click({ force: true });
         cy.get(
           '#pgiea8ec1609e052b7f196935318316299d_1 > :nth-child(1) > .gallery-item-wrapper > :nth-child(1) > a > .gallery-item-content > .gallery-item-visible',