+++ packages/editor/web/src/RichContentEditor/Toolbars/InlineToolbar/InlineToolbar.jsx
@@ -162,8 +162,7 @@ export default class InlineToolbar extends Component {
       isVisible = visibilityFn(editorState);
     }
 
-    // TODO: Test readonly mode and possibly set isVisible to false if the editor is readonly
-    return isVisible || overrideContent || extendContent || keepOpen;
+    return isVisible || overrideContent || extendContent || keepOpen || false;
   };
 
   isVisible = () => this.state.isVisible;