+++ packages/plugin-html/web/src/createHtmlPlugin.js
@@ -4,11 +4,10 @@ import { Component, defaults } from './HtmlComponent';
 import { HTML_TYPE } from './types';
 
 const createHtmlPlugin = (config = {}) => {
-  const type = HTML_TYPE;
-  const { helpers, isMobile, t, [type]: settings = {}, getEditorBounds, ...rest } = config;
+  const { helpers, isMobile, t, [HTML_TYPE]: settings = {}, getEditorBounds, ...rest } = config;
 
   const simulateEditClick = ({ e, pubsub }) =>
-    pubsub.set('onClickTrigger', { event: e, key: 'edit' });
+    setTimeout(() => pubsub.set('onClickTrigger', { event: e, key: 'edit' })); //setTimeout to wait for toolbar to load
 
   return createBasePlugin({
     onOverlayClick: simulateEditClick,