+++ CHANGELOG.md
@@ -20,11 +20,15 @@
 </details>
 <hr/>
 
-## 7.16.2 (Aug 4, 2020)
+## 7.16.3 (Aug 4, 2020)
 ### :bug: Bug Fix
 - `general`
   - [#1413](https://github.com/wix-incubator/rich-content/pull/1413) improve performance - mainly disable the rendering of toolbars when they are not needed
   - [#1413](https://github.com/wix-incubator/rich-content/pull/1413) fix many react warnings for keys  
+
+## 7.16.2 (Aug 4, 2020)
+### :bug: Bug Fix
+- `general`
   - [#1420](https://github.com/wix-incubator/rich-content/pull/1420) added lib entry points to to babel transpilation
 - `common`
   - [#1415](https://github.com/wix-incubator/rich-content/pull/1415) tooltip setstate on an unmounted component and ssr hydration fixed